Fixes #119 some more. Refactored parameter database classes. Fixed Parameter.reinterpolate(), and added Paramater.lookup() as an alternative. Added --xfer-from option to Jones term for explicit interpolation. Added chunk label to Jones machines, for prettier log messages. Minor code cleanups.
