+++ cubical/DefaultParset.cfg
@@ -107,13 +107,16 @@ _Help            = Options for baseline-based corrections (a.k.a. BBCs, a.k.a. i
 load-from        =              # Load and apply BBCs computed in a previous run. Apply with care! This will 
                                   tend to suppress all unmodelled flux towards the centre of the field.
                                   #metavar:FILENAME
+compute-2x2      = 0            # Compute full 2x2 BBCs (as opposed to diagonal-only). Only useful if you
+                                  really trust the polarisation information in your sky model. #type:bool
 apply-2x2        = 0            # Apply full 2x2 BBCs (as opposed to diagonal-only). Only enable this if you
                                   really trust the polarisation information in your sky model. #type:bool
-save-to          = {data[ms]}/BBC-field:{sel[field]}-ddid:{sel[ddid]} # Compute suggested BBCs at end of run,
+save-to          = {data[ms]}/BBC-field:{sel[field]}-ddid:{sel[ddid]}.parmdb # Compute suggested BBCs at end of run,
                                   and save them to the given database. It can be useful to have this always
                                   enabled, since the BBCs provide useful diagnostics of the solution quality 
                                   (and are not actually applied without a load-from setting). #metavar:FILENAME
 per-chan         = 1            # Compute BBCs per-channel (else across entire band). #type:bool
+plot             = 1            # Generate output BBC plots. #type:bool
 
 [dist]
 _Help           = Parallelization and distribution options
@@ -175,16 +178,24 @@ _ExpandedFrom  = --sol-jones    # Identifies --section-option that determines wh
                                   instantiate
 _OtherTemplates = _Help:label   # Colon-separated list of options which need to have their values templated.
 label       = {LABEL}           # Jones label ("G", "dE"). Substituted automatically. #no_cmdline:1 #no_print:1
-solvable    = 1                 # Used in conjunction with -load-from. A non-solvable term is loaded from disk.
-                                  Not to be confused with --sol-jones, which determines the active Jones terms.
+solvable    = 1                 # Set to 0 (and specify -load-from or -xfer-from) to load a non-solvable
+                                  term is loaded from disk. Not to be confused with --sol-jones, which
+                                  determines the active Jones terms.
                                   #type:bool
 type        = complex-2x2       # Type of Jones matrix to solve for. Note that if multiple Jones terms are
                                   enabled, then only complex-2x2 is supported.
                                   #options:complex-2x2|phase-diag|robust-2x2|f-slope|t-slope|tf-plane
-load-from   =                   # Load solutions from given database. By default, the Jones matrix label is
-                                  used to form up parameter names, but his may be overridden by adding
-                                  an explicit "//LABEL" to the database filename. #metavar:FILENAME[//LABEL]
-save-to     = {data[ms]}/{JONES}-field:{sel[field]}-ddid:{sel[ddid]} # Save solutions to given database.
+load-from   =                   # Load solutions from given database. The DB must define solutions
+                                  on the same time/frequency grid (i.e. should normally come from
+                                  calibrating the same pointing/observation). By default, the Jones
+                                  matrix label is used to form up parameter names, but his may be
+                                  overridden by adding an explicit "//LABEL" to the database filename.
+                                  #metavar:FILENAME[//LABEL]
+xfer-from   =                   # Transfer solutions from given database. Similar to -load-from, but
+                                  solutions will be interpolated onto the required time/frequency grid,
+                                  so they can originate from a different field (e.g. from a calibrator).
+                                  #metavar:FILENAME[//LABEL]
+save-to     = {data[ms]}/{JONES}-field:{sel[field]}-ddid:{sel[ddid]}.parmdb # Save solutions to given database.
                                   #metavar:FILENAME
 dd-term     = 0                 # Determines whether this term is direction dependent. --model-ddes must
                                   be enabled. #type:bool
