+++ cubical/machines/complex_2x2_machine.py
@@ -10,7 +10,7 @@ class Complex2x2Gains(PerIntervalGains):
     """
     This class implements the full complex 2x2 gain machine.
     """
-    def __init__(self, label, data_arr, ndir, nmod, chunk_ts, chunk_fs, options):
+    def __init__(self, label, data_arr, ndir, nmod, chunk_ts, chunk_fs, chunk_label, options):
         """
         Initialises a 2x2 complex gain machine.
         
@@ -32,7 +32,8 @@ class Complex2x2Gains(PerIntervalGains):
                 Dictionary of options. 
         """
 
-        PerIntervalGains.__init__(self, label, data_arr, ndir, nmod, chunk_ts, chunk_fs, options)
+        PerIntervalGains.__init__(self, label, data_arr, ndir, nmod,
+                                  chunk_ts, chunk_fs, chunk_label, options)
         
         self.gains     = np.empty(self.gain_shape, dtype=self.dtype)
         self.gains[:]  = np.eye(self.n_cor)
