+++ cubical/main.py
@@ -414,9 +414,9 @@ def main(debugging=False):
         print>>log, ModColor.Str("Exiting with exception: {}({})\n {}".format(type(exc).__name__, 
                                                                     exc, traceback.format_exc()))
         if enable_pdb and not type(exc) is UserInputError:
-            import pdb
+            from cubical.tools import pdb
             exc, value, tb = sys.exc_info()
-            pdb.post_mortem(tb)  # more "modern"
+            pdb.post_mortem(tb)
         sys.exit(2 if type(exc) is UserInputError else 1)
 
 def _io_handler(save=None, load=None, load_model=True, finalize=False):
