+++ cubical/tools/__init__.py
@@ -2,3 +2,10 @@
 # (c) 2017 Rhodes University & Jonathan S. Kenyon
 # http://github.com/ratt-ru/CubiCal
 # This code is distributed under the terms of GPLv2, see LICENSE.md for details
+
+try:
+    import ipdb as pdb
+except ImportError:
+    import pdb
+
+BREAK = pdb.set_trace
